requirements.txt
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
flask>=2.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pytest>=6.0.0
psycopg2-binary>=2.9.0
